update to 0.5.11: some useful bug fixes (check ChangeLog)
Submitted by:	scf
MFC after:	3 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@187791 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ contrib/wpa_supplicant/eap_gpsk_common.c
@@ -376,8 +376,8 @@ static int eap_gpsk_compute_mic_aes(const u8 *sk, size_t sk_len,
 				    const u8 *data, size_t len, u8 *mic)
 {
 	if (sk_len != 16) {
-		wpa_printf(MSG_DEBUG, "EAP-GPSK: Invalid SK length %d for "
-			   "AES-CMAC MIC", sk_len);
+		wpa_printf(MSG_DEBUG, "EAP-GPSK: Invalid SK length %lu for "
+			   "AES-CMAC MIC", (unsigned long) sk_len);
 		return -1;
 	}
